requirements.txt
cobs==1.1.4
future==0.18.2
iso8601==1.0.2
nanopb==0.4.5.post1
protobuf==4.21.2
pycrc==0.9.2
pyserial==3.5
PyYAML==6.0
yapf==0.32.0
pandas
